+++ pref/CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## 1.3.1
+
+- Add stream example
+
 ## 1.3.0
 
 - Implement Stream on key change